Equal files and get_pid fixes.

+++ src/utils/log.cpp
@@ -55,8 +55,8 @@ namespace // unnanmed
 
 static std::string get_pid()
 {
-    char buff[8]={"\0"};
-    sprintf(buff,"%d",getpid());
+    char buff[8]={'\0'};
+    sprintf(buff,"%ld",(long)getpid());
     return std::string(buff);
 }
